// apps/hermitboard-mobile-interface/src/graphql/generated/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /**
   * Define a Relay Cursor type:
   * https://relay.dev/graphql/connections.htm#sec-Cursor
   */
  Cursor: any;
  SessionToken: any;
  Time: any;
  Void: any;
};
// ...
export type CreateConnectionInput = {
  accessToken: Scalars['String'];
  name: Scalars['String'];
  portfolioIDs?: InputMaybe<Array<Scalars['ID']>>;
  refreshToken?: InputMaybe<Scalars['String']>;
};